fix transaction api call placement
