+++ transactions/index.ts
@@ -1,5 +1,5 @@
 import { signBtcTransaction } from './btc';
-import { deDuplicatePendingTx, getNewNonce } from './helper';
+import {getNewNonce } from './helper';
 import {
   addressToString,
   broadcastSignedTransaction,
@@ -8,14 +8,9 @@ import {
   setNonce,
   getNonce,
   setFee,
-  getTransaction,
-  getConfirmedTransactions,
-  fetchStxPendingTxData,
   generateUnsignedStxTokenTransferTransaction,
   estimateFees,
   generateUnsignedSTXTokenTransfer,
-  getTransferTransactions,
-  getMempoolTransactions,
   generateUnsignedTransaction,
   estimateContractCallFees,
   generateUnsignedContractCall,
@@ -30,16 +25,10 @@ export {
   setNonce,
   getNonce,
   setFee,
-  getTransaction,
-  getConfirmedTransactions,
-  getMempoolTransactions,
-  getTransferTransactions,
   generateUnsignedSTXTokenTransfer,
   estimateFees,
   generateUnsignedStxTokenTransferTransaction,
-  fetchStxPendingTxData,
   getNewNonce,
-  deDuplicatePendingTx,
   generateUnsignedTransaction,
   estimateContractCallFees,
   generateUnsignedContractCall,
